Revert "merge vda branch"
