+++ src/main/java/org/red5/server/stream/PlayEngine.java
@@ -163,7 +163,7 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
     /**
      * State machine for video frame dropping in live streams
      */
-    private final IFrameDropper frameDropper = new FrameDropper();
+    private IFrameDropper videoFrameDropper = new VideoFrameDropper();
 
     private int timestampOffset = 0;
 
@@ -225,7 +225,7 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
     /**
      * List of pending operations
      */
-    private final ConcurrentLinkedQueue<Runnable> pendingOperations = new ConcurrentLinkedQueue<Runnable>();
+    private ConcurrentLinkedQueue<Runnable> pendingOperations = new ConcurrentLinkedQueue<Runnable>();
 
     // Keep count of dropped packets so we can log every so often.
     private long droppedPacketsCount = 0;
@@ -412,7 +412,7 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                 in = providerService.getLiveProviderInput(thisScope, itemName, false);
                 if (msgInReference.compareAndSet(null, in)) {
                     // drop all frames up to the next keyframe
-                    frameDropper.reset(IFrameDropper.SEND_KEYFRAMES_CHECK);
+                    videoFrameDropper.reset(IFrameDropper.SEND_KEYFRAMES_CHECK);
                     if (in instanceof IBroadcastScope) {
                         IBroadcastStream stream = (IBroadcastStream) ((IBroadcastScope) in).getClientBroadcastStream();
                         if (stream != null && stream.getCodecInfo() != null) {
@@ -426,15 +426,15 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                                 sendNotifications = false;
                                 if (videoCodec.getNumInterframes() > 0 || videoCodec.getKeyframe() != null) {
                                     bufferedInterframeIdx = 0;
-                                    frameDropper.reset(IFrameDropper.SEND_BUFFERED_INTERFRAMES);
+                                    videoFrameDropper.reset(IFrameDropper.SEND_ALL);
                                 }
                             }
                         }
                     }
                     // subscribe to stream (ClientBroadcastStream.onPipeConnectionEvent)
+                    in.subscribe(this, null);
                     // execute the processes to get Live playback setup
                     playLive();
-                    in.subscribe(this, null);
                 } else {
                     sendStreamNotFoundStatus(currentItem);
                     throw new StreamNotFoundException(itemName);
@@ -762,7 +762,7 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                     }
                 } else {
                     subscriberStream.onChange(StreamState.RESUMED, currentItem, position);
-                    frameDropper.reset(IFrameDropper.SEND_KEYFRAMES_CHECK);
+                    videoFrameDropper.reset(VideoFrameDropper.SEND_KEYFRAMES_CHECK);
                 }
                 break;
             default:
@@ -1062,7 +1062,7 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
         // if the connection is congested for LIVE streams.
         RTMPMessage messageOut = RTMPMessage.build(event, messageIn.getBody().getSourceType());
         //get the current timestamp from the message
-        int ts = messageOut.getBody().getTimestamp() - frameDropper.getDroppedGapDuration();
+        int ts = messageOut.getBody().getTimestamp();
         if (log.isTraceEnabled()) {
             log.trace("sendMessage: streamStartTS={}, length={}, streamOffset={}, timestamp={}", new Object[] { streamStartTS.get(), currentItem.getLength(), streamOffset, ts });
             final long delta = System.currentTimeMillis() - playbackStart;
@@ -1484,7 +1484,7 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                         log.info("Dropping packet because we are paused. sessionId={} stream={} count={}",
                                 sessionId, subscriberStream.getBroadcastStreamPublishName(), droppedPacketsCount);
                     }
-                    frameDropper.dropPacket(rtmpMessage);
+                    videoFrameDropper.dropPacket(rtmpMessage);
                     return;
                 }
 
@@ -1498,7 +1498,7 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                             // dont try to drop frames if video codec is null
                             if (videoCodec != null && videoCodec.canDropFrames()) {
                                 if (!receiveVideo) {
-                                    frameDropper.dropPacket(rtmpMessage);
+                                    videoFrameDropper.dropPacket(rtmpMessage);
                                     droppedPacketsCount++;
                                     if (log.isInfoEnabled() && shouldLogPacketDrop()) {
                                         // client disabled video or the app doesn't have enough bandwidth allowed for this stream
@@ -1525,20 +1525,15 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                                                     droppedPacketsCount});
                                 }
 
-                                if (!frameDropper.canSendPacket(rtmpMessage, pendingVideos)) {
+                                if (!videoFrameDropper.canSendPacket(rtmpMessage, pendingVideos)) {
                                     // drop frame as it depends on other frames that were dropped before
                                     droppedPacketsCount++;
-                                    if (pendingVideos != 0) {
-                                        frameDropper.skipPacket(rtmpMessage);
-                                        return;
-                                    }
                                     if (log.isInfoEnabled() && shouldLogPacketDrop()) {
                                         log.info("Frame dropper says to drop packet. sessionId={} stream={} count={}",
                                                 sessionId, subscriberStream.getBroadcastStreamPublishName(), droppedPacketsCount);
                                     }
-                                    frameDropper.reset(IFrameDropper.SEND_BUFFERED_INTERFRAMES);
                                     return;
-                                } else {
+                                }
                                 // increment the number of times we had pending video frames sequentially
                                 if (pendingVideos > 1) {
                                     numSequentialPendingVideoFrames++;
@@ -1562,43 +1557,21 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                                         sendInsufficientBandwidthStatus(currentItem);
                                         nextCheckBufferUnderrun = now + bufferCheckInterval;
                                     }
-                                        frameDropper.dropPacket(rtmpMessage);
+                                    videoFrameDropper.dropPacket(rtmpMessage);
                                     return;
                                 }
-                                }
-
-                                int lastDropTimestamp = frameDropper.getLastDropTimestamp();
-                                /* check if we have buffered frame set newer than lastDropTimestamp.
-                                /* if so - we should start from buffered keyframe and update frameDropper*/
-                                if (frameDropper.hasDroppedPackets() &&
-                                        videoCodec.getKeyframeTimestamp() > lastDropTimestamp) {
-                                    frameDropper.updateLastDropTimestamp(videoCodec.getKeyframeTimestamp());
-                                    frameDropper.reset(IFrameDropper.SEND_BUFFERED_KEYFRAME);
-                                }
-
                                 // we are ok to send, check if we should send buffered frame
-                                if (frameDropper.getState() == IFrameDropper.SEND_BUFFERED_INTERFRAMES) {
-                                    if (frameDropper.getBufferedInterframeIdx() >= videoCodec.getNumInterframes()) {
-                                        frameDropper.reset(IFrameDropper.SEND_ALL);
+                                if (bufferedInterframeIdx > -1) {
+                                    IVideoStreamCodec.FrameData fd = videoCodec.getInterframe(bufferedInterframeIdx++);
+                                    if (fd != null) {
+                                        VideoData interframe = new VideoData(fd.getFrame());
+                                        interframe.setTimestamp(body.getTimestamp());
+                                        rtmpMessage = RTMPMessage.build(interframe);
                                     } else {
-                                        IVideoStreamCodec.FrameData fd;
-                                        //skip to interframe later then lastDropTimestamp
-                                        do {
-                                            fd = videoCodec.getInterframe(frameDropper.getAndIncrementBufferedInterframeIdx());
-                                        } while (fd.getTimestamp() <= lastDropTimestamp &&
-                                                frameDropper.getBufferedInterframeIdx() <= videoCodec.getNumInterframes());
-                                        VideoData bufferedFrame = new VideoData(fd.getFrame());
-                                        bufferedFrame.setTimestamp(body.getTimestamp());
-                                        rtmpMessage = RTMPMessage.build(bufferedFrame);
-                                        fd = null;
+                                        // it means that new keyframe was received and we should send current frames instead of buffered
+                                        bufferedInterframeIdx = 0;
                                     }
-                                } else if (frameDropper.getState() == IFrameDropper.SEND_BUFFERED_KEYFRAME) {
-                                    VideoData bufferedFrame = new VideoData(videoCodec.getKeyframe());
-                                    bufferedFrame.setTimestamp(body.getTimestamp());
-                                    rtmpMessage = RTMPMessage.build(bufferedFrame);
-                                    frameDropper.reset(IFrameDropper.SEND_BUFFERED_INTERFRAMES);
                                 }
-                                frameDropper.sendPacket(rtmpMessage);
                             }
                         }
                     }
@@ -1614,11 +1587,6 @@ public final class PlayEngine implements IFilter, IPushableConsumer, IPipeConnec
                         }
                         rtmpMessage = RTMPMessage.build(body);
                     } else if (!receiveAudio) {
-                        return;
-                    }
-                    if (!frameDropper.canSendPacket(rtmpMessage, 0)) {
-                        // Drop frame as it depends on other frames that were dropped before.
-                        log.debug("Dropping audio packet because frame dropper says we cant send it");
                         return;
                     }
                 }
