+++ src/main/java/org/red5/server/stream/PlaylistSubscriberStream.java
@@ -36,7 +36,6 @@ import org.red5.server.api.stream.IPlayItem;
 import org.red5.server.api.stream.IPlaylistController;
 import org.red5.server.api.stream.IPlaylistSubscriberStream;
 import org.red5.server.api.stream.IStreamAwareScopeHandler;
-import org.red5.server.api.stream.IStreamCapableConnection;
 import org.red5.server.api.stream.OperationNotSupportedException;
 import org.red5.server.api.stream.StreamState;
 import org.slf4j.Logger;
@@ -173,50 +172,44 @@ public class PlaylistSubscriberStream extends AbstractClientStream implements IP
 
     /** {@inheritDoc} */
     public void start() {
-        write.lock();
-        try {
-            IStreamCapableConnection conn = getConnection();
-            //ensure the play engine exists
-            if (engine == null && conn != null && conn.isConnected()) {
-                IScope scope = getScope();
-                if (scope != null) {
-                    IContext ctx = scope.getContext();
-                    if (ctx.hasBean(ISchedulingService.BEAN_NAME)) {
-                        schedulingService = (ISchedulingService) ctx.getBean(ISchedulingService.BEAN_NAME);
-                    } else {
-                        //try the parent
-                        schedulingService = (ISchedulingService) scope.getParent().getContext().getBean(ISchedulingService.BEAN_NAME);
-                    }
-                    IConsumerService consumerService = null;
-                    if (ctx.hasBean(IConsumerService.KEY)) {
-                        consumerService = (IConsumerService) ctx.getBean(IConsumerService.KEY);
-                    } else {
-                        //try the parent
-                        consumerService = (IConsumerService) scope.getParent().getContext().getBean(IConsumerService.KEY);
-                    }
-                    IProviderService providerService = null;
-                    if (ctx.hasBean(IProviderService.BEAN_NAME)) {
-                        providerService = (IProviderService) ctx.getBean(IProviderService.BEAN_NAME);
-                    } else {
-                        //try the parent
-                        providerService = (IProviderService) scope.getParent().getContext().getBean(IProviderService.BEAN_NAME);
-                    }
-                    engine = new PlayEngine.Builder(this, schedulingService, consumerService, providerService).build();
+        //ensure the play engine exists
+        if (engine == null) {
+            IScope scope = getScope();
+            if (scope != null) {
+                IContext ctx = scope.getContext();
+                if (ctx.hasBean(ISchedulingService.BEAN_NAME)) {
+                    schedulingService = (ISchedulingService) ctx.getBean(ISchedulingService.BEAN_NAME);
                 } else {
-                    throw new IllegalStateException("Scope was null on start playing");
+                    //try the parent
+                    schedulingService = (ISchedulingService) scope.getParent().getContext().getBean(ISchedulingService.BEAN_NAME);
                 }
+                IConsumerService consumerService = null;
+                if (ctx.hasBean(IConsumerService.KEY)) {
+                    consumerService = (IConsumerService) ctx.getBean(IConsumerService.KEY);
+                } else {
+                    //try the parent
+                    consumerService = (IConsumerService) scope.getParent().getContext().getBean(IConsumerService.KEY);
+                }
+                IProviderService providerService = null;
+                if (ctx.hasBean(IProviderService.BEAN_NAME)) {
+                    providerService = (IProviderService) ctx.getBean(IProviderService.BEAN_NAME);
+                } else {
+                    //try the parent
+                    providerService = (IProviderService) scope.getParent().getContext().getBean(IProviderService.BEAN_NAME);
+                }
+                engine = new PlayEngine.Builder(this, schedulingService, consumerService, providerService).build();
+            } else {
+                throw new IllegalStateException("Scope was null on start playing");
             }
-            //set buffer check interval
-            engine.setBufferCheckInterval(bufferCheckInterval);
-            //set underrun trigger
-            engine.setUnderrunTrigger(underrunTrigger);
-            // Start playback engine
-            engine.start();
-            // Notify subscribers on start
-            onChange(StreamState.STARTED);
-        } finally {
-            write.unlock();
         }
+        //set buffer check interval
+        engine.setBufferCheckInterval(bufferCheckInterval);
+        //set underrun trigger
+        engine.setUnderrunTrigger(underrunTrigger);
+        // Start playback engine
+        engine.start();
+        // Notify subscribers on start
+        onChange(StreamState.STARTED);
     }
 
     /** {@inheritDoc} */
@@ -304,23 +297,18 @@ public class PlaylistSubscriberStream extends AbstractClientStream implements IP
         if (log.isDebugEnabled()) {
             log.debug("close");
         }
-        write.lock();
-        try {
-            if (engine != null) {
-                // before or on close we may need to allow the queued messages a chance to clear
-                engine.close();
-                onChange(StreamState.CLOSED);
-                items.clear();
-                // clear jobs
-                if (schedulingService != null && !jobs.isEmpty()) {
-                    for (String jobName : jobs) {
-                        schedulingService.removeScheduledJob(jobName);
-                    }
-                    jobs.clear();
+        if (engine != null) {
+            // before or on close we may need to allow the queued messages a chance to clear
+            engine.close();
+            onChange(StreamState.CLOSED);
+            items.clear();
+            // clear jobs
+            if (schedulingService != null && !jobs.isEmpty()) {
+                for (String jobName : jobs) {
+                    schedulingService.removeScheduledJob(jobName);
                 }
+                jobs.clear();
             }
-        } finally {
-            write.unlock();
         }
     }
